backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["louis <louisjoejordan@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.scripts]
start = "src.main:start"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
elevenlabs = "^1.9.0"
python-dotenv = "^1.0.1"
websockets = "^13.0.1"
openai = "^1.46.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
